Move portalsearchpage translation tests to ask_cfpb
